+++ NEWS.md
@@ -48,6 +48,7 @@ A major refactor went into the codebase to cleanup several leaking processes. Th
 - You can now change the default applications from the settings
 - Added a new battery plugin for the settings
 - Handle dpi better for smaller displays ( < 1920x1080)
+- You can now tab-complete some parts in the prompt
 
 ### Bug Fixes
 - Don't play a 'pop' sound when opening the left-panel